+++ lab4/fgjyh/go.mod
@@ -1,0 +1,3 @@
+module cal
+
+go 1.17
